Rename Dev views
