+++ AvaloniaUIClient/Infrastructure/Views/DevPanelViews/LogsPanel/LogsDevSettingsView.axaml
@@ -1,5 +1,5 @@
 <UserControl
-    x:Class="AvaloniaUIClient.Infrastructure.Views.DevPanelViews.LogsPanel.LogsSettingsView"
+    x:Class="AvaloniaUIClient.Infrastructure.Views.DevPanelViews.LogsPanel.LogsDevSettingsView"
     xmlns="https://github.com/avaloniaui"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
