+++ AvaloniaUIClient/Infrastructure/Views/DevPanelViews/LogsPanel/LogsDevView.axaml
@@ -1,5 +1,5 @@
 ﻿<UserControl
-    x:Class="AvaloniaUIClient.Infrastructure.Views.DevPanelViews.LogsPanel.LogsPanelView"
+    x:Class="AvaloniaUIClient.Infrastructure.Views.DevPanelViews.LogsPanel.LogsDevView"
     xmlns="https://github.com/avaloniaui"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:avalonia="clr-namespace:Material.Icons.Avalonia;assembly=Material.Icons.Avalonia"
