+++ AvaloniaUIClient/Infrastructure/Views/DevPanelViews/LogsPanel/LogsDevView.axaml.cs
@@ -1,11 +1,11 @@
 ﻿using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace AvaloniaUIClient.Infrastructure.Views.DevPanelViews;
+namespace AvaloniaUIClient.Infrastructure.Views.DevPanelViews.LogsPanel;
 
-public partial class DevPanelView : UserControl
+public partial class LogsDevView : UserControl
 {
-    public DevPanelView()
+    public LogsDevView()
     {
         InitializeComponent();
     }
