--- AvaloniaUIClient/Infrastructure/Views/DevPanelViews/LogsPanel/LogsPanelView.axaml.cs
@@ -1,17 +1,0 @@
-﻿using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
-
-namespace AvaloniaUIClient.Infrastructure.Views.DevPanelViews.LogsPanel;
-
-public partial class LogsPanelView : UserControl
-{
-    public LogsPanelView()
-    {
-        InitializeComponent();
-    }
-
-    private void InitializeComponent()
-    {
-        AvaloniaXamlLoader.Load(this);
-    }
-}
