+++ AvaloniaUIClient/Infrastructure/Views/DevPanelViews/StoresDevView.axaml.cs
@@ -1,0 +1,17 @@
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace AvaloniaUIClient.Infrastructure.Views.DevPanelViews;
+
+public partial class StoresDevView : UserControl
+{
+    public StoresDevView()
+    {
+        InitializeComponent();
+    }
+
+    private void InitializeComponent()
+    {
+        AvaloniaXamlLoader.Load(this);
+    }
+}
