+++ AvaloniaUIClient/Infrastructure/Views/DevPanelViews/StoresDevlView.axaml
@@ -1,5 +1,5 @@
 <UserControl
-    x:Class="AvaloniaUIClient.Infrastructure.Views.DevPanelViews.StoresPanelView"
+    x:Class="AvaloniaUIClient.Infrastructure.Views.DevPanelViews.StoresDevView"
     xmlns="https://github.com/avaloniaui"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
